+++ projects/ids-enterprise-ng/src/lib/rating/soho-rating.spec.ts
@@ -29,7 +29,7 @@ describe('Soho Rating Unit Tests', () => {
 });
 
 @Component({
-  template: `<div soho-rating>`
+  template: `<div soho-rating></div>`
 })
 class SohoRatingTestComponent {
   @ViewChild(SohoRatingComponent) rating?: SohoRatingComponent;
@@ -39,6 +39,7 @@ describe('Soho Rating Chart Render', () => {
   let fixture: ComponentFixture<SohoRatingTestComponent>;
   let de: DebugElement;
   let el: HTMLElement;
+  let component: SohoRatingComponent | undefined;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -50,10 +51,21 @@ describe('Soho Rating Chart Render', () => {
 
     de = fixture.debugElement;
     el = de.query(By.css('[soho-rating]')).nativeElement;
+    component = fixture.componentInstance.rating;
   });
 
   it('Check HTML content', () => {
     expect(el.hasAttribute('soho-rating')).toBeTruthy('soho-rating');
+  });
+
+  it ('Check readonly/enable', () => {
+    if (component) {
+      component.readonly();
+      expect(el.classList.contains('is-readonly')).toBeTruthy();
+
+      component.enable();
+      expect(el.classList.contains('is-readonly')).toBeFalsy();
+    }
   });
 
 });